Respect Leading Zeros in Initial Render Pass
Currently, an element that is initialized while having a textContent of
"000", will be shown as "0". This commit takes a snapshop of the initial
value, and then uses that snapshot in the initial render pass, rather
than converting the string to a number and then back to a string. This
respects leading zeros.